refactor: separate UI and logic from old `CategoryCreationDrawer`
