+++ src/components/ui/ShareHouseManagementHead.tsx
@@ -2,7 +2,7 @@
 import { useState } from 'react';
 import { CirclePlus } from 'lucide-react';
 import { TenantInvitationDrawer } from '@/components/ui/drawers/tenants/TenantInvitationDrawer';
-import { CategoryCreationDrawer } from './drawers/CategoryCreationDrawer';
+import { CategoryCreationDrawer } from './drawers/categories/CategoryCreationDrawer';
 
 interface IHeaderTitleButtonProps {
   shareHouseId: string;
@@ -46,8 +46,8 @@ export const ShareHouseManagementHead = ({
       <div className="h-0">
         <CategoryCreationDrawer
           shareHouseId={shareHouseId}
-          open={openTaskDrawer}
-          setOpen={setOpenTaskDrawer}
+          editOpen={openTaskDrawer}
+          setEditOpen={setOpenTaskDrawer}
         />
         <TenantInvitationDrawer
           shareHouseId={shareHouseId}
