+++ src/components/ui/drawers/task-description/TaskDescriptionRenderer.tsx
@@ -1,7 +1,7 @@
 import { EditorContent, useEditor } from '@tiptap/react';
-import { editorExtensions } from '../CategoryCreationDrawer';
-import { useFormContext } from 'react-hook-form';
+import { editorExtensions } from '@/components/ui/drawers/categories/CategoryDrawerContent';
 import { TTaskSchema } from '@/components/ui/drawers/tasks/TaskDrawerContent';
+import { useFormContext } from 'react-hook-form';
 
 // Difference between TaskDescriptionRendererForCategory.tsx and TaskDescriptionEditor.tsx is only the schema type used
 export const TaskDescriptionRenderer = () => {
